Moved test walker
